Autoloader: check the parent directory permissions before trying to delete files.
Verify that the parent directory is writable and executable before trying to delete files.

+++ packages/autoloader/src/class-autoloader-handler.php
@@ -85,7 +85,7 @@ class Autoloader_Handler {
 
 	/**
 	 * Check the plugins directory for autoloader v1.x files. If the plugin's directory contains
-	 * v1.x autoloader files, we will ignore it.
+	 * v1.x autoloader files, we will ignore this plugins autoloader.
 	 *
 	 * This situation can occur if a plugin has a v2.x autoloader and a backup containing a version
 	 * of the plugin with a v1.x autoloader is restored. The v2.x autolaoder files remain in the
@@ -98,32 +98,49 @@ class Autoloader_Handler {
 	private function check_for_v1_files( $autoload_packages_path ) {
 		$v1_classmap_path = trailingslashit( dirname( $autoload_packages_path ) ) . 'composer/autoload_classmap_package.php';
 
+		clearstatcache( true, $v1_classmap_path );
+
 		if ( ! file_exists( $v1_classmap_path ) ) {
 			return false;
 		}
 
-		clearstatcache();
+		clearstatcache( true, $autoload_packages_path );
 
-		// The v2.x file is small, less than 1 KB. The v1.x file is much larger, > 6 KB.
-		if ( 1000 < filesize( $autoload_packages_path ) ) {
-			// Try to delete v2.x files.
-			if ( file_exists( trailingslashit( dirname( $autoload_packages_path ) ) . 'autoload_functions.php' ) ) {
-				unlink( trailingslashit( dirname( $autoload_packages_path ) ) . 'autoload_functions.php' );
+		/*
+		 * Use the filesize of autoload_packages.php to determine if it's a v1 or v2 file.
+		 * The v2.x file is small, less than 1 KB. The v1.x file is much larger, > 6 KB.
+		 */
+		if ( 2000 < filesize( $autoload_packages_path ) ) {
+			// The autoload_package.php file is a v1.x, so try to delete v2.x files.
+			$vendor_dir = dirname( $autoload_packages_path );
+			if ( file_exists( trailingslashit( $vendor_dir ) . 'autoload_functions.php' )
+				&& is_writable( $vendor_dir )
+				&& is_executable( $vendor_dir ) ) {
+				unlink( trailingslashit( $vendor_dir ) . 'autoload_functions.php' );
 			}
 
-			if ( file_exists( trailingslashit( dirname( $autoload_packages_path ) ) . 'jetpack-autoloader/autoload_functions.php' ) ) {
-				unlink( trailingslashit( dirname( $autoload_packages_path ) ) . 'jetpack-autoloader/autoload_functions.php' );
+			$jetpack_autoloader_dir = trailingslashit( $vendor_dir ) . 'jetpack-autoloader';
+			if ( file_exists( trailingslashit( $jetpack_autoloader_dir ) . 'autoload_functions.php' )
+				&& is_writable( $jetpack_autoloader_dir )
+				&& is_executable( $jetpack_autoloader_dir ) ) {
+				unlink( trailingslashit( $jetpack_autoloader_dir ) . 'autoload_functions.php' );
 			}
 
-			if ( file_exists( trailingslashit( dirname( $autoload_packages_path ) ) . 'composer/jetpack_autoload_classmap.php' ) ) {
-				unlink( trailingslashit( dirname( $autoload_packages_path ) ) . 'composer/jetpack_autoload_classmap.php' );
+			$composer_dir = trailingslashit( $vendor_dir ) . 'composer';
+			if ( file_exists( trailingslashit( $composer_dir ) . 'jetpack_autoload_classmap.php' )
+				&& is_writable( $composer_dir )
+				&& is_executable( $composer_dir ) ) {
+				unlink( trailingslashit( $composer_dir ) . 'jetpack_autoload_classmap.php' );
 			}
 
 			return true;
 		}
 
 		// Try to delete v1.x classmap file.
-		unlink( $v1_classmap_path );
+		if ( is_writable( dirname( $v1_classmap_path ) ) && is_executable( dirname( $v1_classmap_path ) ) ) {
+			unlink( $v1_classmap_path );
+		}
+
 		return false;
 	}
